Run tests in CI using github actions (#52)
## Problem

This repository has some tests, but they are not being run in CI. We
need to setup this integration so we can evaluate the impact of PRs and
begin to collaborate with more people.

## Solution

Implement a github actions workflow to run tests.

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [x] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

Observe things running in Github Actions.

## Notes

- There's definitely some kind of test timing issue causing these tests
to flake. But some tests is better than no tests.