[v1] Allow createIndex to support waitUntilReady for async polling (#89)
## Problem
Creating a new index is an asynchronous operation on the backend. When
calling `client.createIndex({ ... })` the operation to trigger creating
the index occurs nearly immediately, but the index will not actually be
ready for some amount of time after the creation call returns (somewhere
between 10-13 seconds average currently).

v0 offered a utility function `waitUntilIndexIsReady` which allowed an
async call leveraging polling `describeIndex()`.

## Solution
- Add a new `waitUntilReady` option to `createIndex()` allowing the user
to opt for returning an async function that polls `describeIndex()`
until the index is ready.
- Add a new `waitUntilIndexIsReady()` function in `/createIndex.ts` to
handle polling `describeIndex()` if desired. Originally I was going to
put the function in `/utils` but it felt specific enough to creating the
index that it made more sense living here for now.
- Update `createIndex.test.ts`
   - Added some error case tests for the basic functionality.
- Added tests to run through the asynchronous logic of
`waitUntilIndexIsReady()`.
- Upgraded `@types/jest` as there was some functionality in Jest 29.5.0
that wasn't represented. Did a lot of troubleshooting around jest and
mocking timers while working on tests for this, they had [exposed a
bunch of new functions recently for working with fake
timers](https://github.com/jestjs/jest/pull/13981) and autocomplete lead
me astray. 😭

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [X] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
Validate that passing `waitUntilReady: true` to `createIndex()` await
the creation of the index on the backend.

```
$ npm run repl

$ await init()

// Wait until index is ready
$ await client.createIndex({ name: 'test-create-1', dimension: 5, waitUntilReady: true })
'Index is ready after 13 seconds'

// Test the basic scenario of returning as soon as the create call completes
$ await client.createIndex({ name: 'test-create-2', dimension: 5 })
undefined
```