[v1] Migrate vector/record terminology and add missing type exports (#96)
## Problem

1) We have adopted the term `Record` in our
[documentation](https://docs.pinecone.io/docs/overview#pinecone-indexes-store-records-with-vector-data)
to refer to items stored in an index.
2) Several type exports were missing from the module.

## Solution

- Eliminate duplication of some type definitions by moving definition
for shared concepts such as `Record` and `Index` into `/types.ts` files
that can be referenced from multiple actions.
- Rename the references to `Vector` to `PineconeRecord` everywhere it
makes sense. Alias values to the name `vectors` when interacting with
the generated client. I chose to avoid using the name `Record` for the
type because there is already a [typescript utility with this
name](https://www.typescriptlang.org/docs/handbook/utility-types.html#recordkeys-type)
and it seems to confuse the typescript compiler and lead to confusing
hints & error messages.
- Add module exports for every client methods argument and return types.
Some of these types such as `DeleteOneOptions` are just aliases of other
types, but giving them a method-specific name affords some flexibility
for refactoring in the future if we decide to extend the method to
accept other things without introducing a breaking type change.

New files:
- `src/data/types.ts`
- `src/control/types.ts`

## Type of Change

- [x] Bug fix (non-breaking change which fixes an issue)
- [x] This change requires a documentation update

## Test Plan

Only types and exported types changed in this diff. Tests should be
passing with changes only to import paths.