[v1] Change fetch return field from `vectors` to `records` (#100)
## Problem

"Record" is now the [preferred
terminology](https://docs.pinecone.io/docs/overview#pinecone-indexes-store-records-with-vector-data)
for items stored in an index. The fetch command still refers to
`vectors` in the response.

## Solution

Change the name of the property in the results object. Update tests.


## Type of Change

- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] This change requires a documentation update

## Test Plan

After the change, the REPL shows:

```typescript
> let pinecone = new Pinecone()
undefined
> await pinecone.index('jen2').fetch(['1'])
{
  records: {
    '1': {
      id: '1',
      values: [Array],
      sparseValues: undefined,
      metadata: [Object]
    }
  },
  namespace: ''
}
```