Add support for Edge Runtime (#110)
## Problem

People attempting to use this library in Vercel's [Edge
Runtime](https://edge-runtime.vercel.app/) are hitting runtime
exceptions due to reliance on some node utilities that are not
available.

This addresses issue #108 

## Solution

- Setup new `npm run test:integration:edge` which runs integration tests
using a special jest context designed to highlight when code is using
dependencies that are not available on edge.
- Integrate this test run into CI

Code changes to get these tests passing:
- Add an `isEdge()` utility that looks for a global var, `EdgeRuntime`
to detect whether execution is happening in the Edge Runtime.
- Use that util to guard things which won't work in Edge. Things not
supported in Edge:
- Pulling `os` info for the user-agent string. This info can just be
omitted from the user-agent.
- Using `fs` to read `package.json` version, also for the user-agent
string. The version number is the most important thing in the user-agent
so I do need some other way of getting that info. Importing json as a
module is possible in es6, but I was getting typescript build errors
when trying to grab package.json directly because that file is outside
of the tsconfig `rootDir`. So, I adjusted the publish steps to dump
version info to a place it can be safely referenced, `src/version.json`.
- Runtime validation using AJV. The code generation that happens in the
`compile` step causes Edge to blow up. So we skip runtime validations
for Edge for now. If they are developing with TypeScript, these should
be largely unnecessary anyway.

## Type of Change

- [x] New feature (non-breaking change which adds functionality)
- [x] Infrastructure change (CI configs, etc)

## Test Plan

- New CI job for edge integration tests should be green.
- For changes to publish jobs, I did some local testing to explore what
the npm version command does and test the jq commands. But won't know
for sure that this is all correct, though, until trying to ship a
release.