Update typedoc.json configurations and type exports (#119)
## Problem
There are some basic configurations for TypeDoc we can provide to clean
up some of the documentation output. There were also a number of
warnings being output when running the utility calling out internals
that are represented in documentation, but not included due to not being
available as an export at the top level.

## Solution

- Update `typedoc.json` config to exclude `/v0/` code paths, include the
version number in the output, and hide the generator tagline at the
bottom of the page.
- Update the `{@link PineconeArgumentError}` doc strings to `{@link
Errors.PineconeArgumentError` to properly pull in the references.
- Export missing types: `IndexName`, `CollectionName`,
`DeleteManyByVectorIdOptions`, `DeleteManyByFilterOptions`, and
`QueryShared`.

## Type of Change
- [X] Infrastructure change (CI configs, etc)

## Test Plan
Make sure CI is green.
Validate documentation after it's been built and deployed:
https://pinecone-io.github.io/pinecone-ts-client/

The documentation CI step should also still show warnings for some of
the `pinecone-generated-ts-fetch` internals, and a few of the index
classes. I'm honestly unsure if those should be exported, and exporting
them at the top level pulled in additional pieces into warnings so it
may need a more thorough pass.

Current warnings should look like this:
<img width="1484" alt="Screenshot 2023-09-14 at 4 39 59 PM"
src="https://github.com/pinecone-io/pinecone-ts-client/assets/119623786/0a9568c3-1a50-40c2-a44b-c73e12aac618">