[Bugfix] Implement next.js fetch compatibility (#126)
## Problem

Next.js seems to stub out the cross-fetch polyfill which has the effect
of silently breaking code which depends on cross-fetch to make API
requests.

Related bugs: 
- #124 
- v1 migration problems in
https://github.com/pinecone-io/pinecone-vercel-starter/pull/14

## Solution

I have added a `getFetch` utility which looks for a `fetch`
implementation already loaded in the global scope and returns the
cross-fetch implementation only as a last resort.

- Now when using our client library in a next.js project, this should
result in the `@vercel/fetch` implementation being used in favor of the
disabled polyfill.
- When using this client library in other environments where next.js is
not being used, it still activates the cross-fetch polyfill.
- Also, I've exposed an experimental configuration option for anyone who
wants to pass in a different fetch implementation to use. This could
open up some flexibility for people with specific network configurations
who need the ability to configure the http client being used in some
way. If the user provides a fetch implementation through the `fetchApi`
config option, it will take priority over a fetch in the global scope or
the cross-fetch polyfill.

### Other changes

- Added additional integration tests for `query` method
- Added an explicit dependency on the `encoding` module to resolve a
warning message [similar to this
one](https://github.com/vercel/next.js/discussions/54109)

## Type of Change

- [x] Bug fix (non-breaking change which fixes an issue)

## Test Plan

I am currently testing this change in the context of work-in-progress to
migrate our vercel-starter example to using the v1 version of the
client. https://github.com/pinecone-io/pinecone-vercel-starter/pull/14

- Checkout the PR branch from pinecone-vercel-starter in a sibling
directory
- Install my local changes to this client with `npm install
../pinecone-ts-client`
- Try to run the starter project with `npm run dev`
- Attempt to use the sample app and observe it progresses beyond the
error related broken cross-fetch