Overhaul error handling to address text() error (#135)
## Problem

Some users are still running into errors that can't be properly
diagnosed because of a bug in error handling causing the real error to
get obscured by an unrelated problem. There's a point at which we assume
we should be dealing with a `ResponseError` with a `text()` field; in
reality, that's not a correct assumption so it's not safe to cast to
that type.

- #124 

## Solution

This was one of those changes where you pull on the thread and the whole
carpet unravels.

- Refactor error handling away from nested callbacks into a flat error
handler function
- Wrap the underlying exception by attaching it as a `cause` whenever a
`PineconeConnectionError` exception is being thrown. This will help us
investigate #124
- Migrate away from method-specific error handling to a consistent
approach based on middleware functions. We already had an
httpErrorMapper, but with this approach it only has to get wired up
twice instead of like 20 times so that really simplifies what we need to
cover in unit testing.
- Implement a bunch of new integration tests to give confidence the
middleware is actually working as expected.
- The cost of this change is we lose a little bit of convenience in the
404 case where we were fetching and presenting the list of valid
indexes/collections. But that small bit of convenience doesn't seem
worth a radically more complex approach to error handling.
- We no longer need to test exception handling in most method
implementations
- It unlocks the ability for some really cool debug options. See #136.

#### Summary of files changed:
- **Error handling logic** in `src/errors/handling.ts`: 
    - Refactor from nested callbacks into a flat error handler function.
- Continue to delegate to httpErrorMapper, but check that the error
actually is a ResponseError before trying to access the `text()`
property.
- For other errors, wrap the underlying errors and throw
`PineconeConnectionError`
- **Error classes** in `src/errors`: 
- `PineconeConnectionError` and `BasePineconeError` modified to allow
reference to underlying errors to be captured.
- Revised error message that explains more reasons why
`PineconeConnectionError` might occur.
- **Middleware**: These get invoked at different points in the lifecycle
of every request.
    - Defined middleware using `handleApiError` 
- Wired up new middleware for error handling in `src/pinecone.ts` and
`src/data/index.ts`. Now we can be confident we are handling errors
consistently across every method without the need for a lot of
repetitive unit testing to verify wiring on each client method.
- **New integration tests**. 
- Added a more robust suite of integration tests to exercise the error
handling in a variety of real scenarios.
    - Deleted method-specific error handling that is no longer needed
    -  Deleted method-level unit testing of error handling
 

## Type of Change

- [x] Bug fix (non-breaking change which fixes an issue)

## Test Plan

Describe specific steps for validating this change.