Refactor projectIdSingleton -> hostUrlSingleton (#148)
## Problem
In the future, we'll be using `describeIndex` to retrieve the specific
`host` value for a given Index. Previously, this was handled internally
by the client via configuration values and pinging `whoami`, and caching
the results to prevent unneeded requests and latency.

We want to maintain this caching behavior for `apiKey` -> `indexName`
while simplifying the client configuration and calling `describeIndex`
to retrieve the `host`.

## Solution
The new `IndexHostSingleton` replaces the existing `ProjectIdSingleton`.
The functionality is very similar between the two singletons at a high
level, but the code paths for making network calls are different. I've
also gone ahead and pulled out `projectId` and `environment` as
configuration values as they'll no longer be needed to generate the data
plane address.

- Delete `ProjectIdSingleton` and add new `IndexHostSingleton`.
- Add unit tests for `IndexHostSingleton`.
- Remove `projectId` and `environment` from `PineconeConfiguration` and
update related areas, tests, documentation bits, etc. Add new
`PINECONE_CONTROLLER_HOST` environment variable.
- Add `.then()` callbacks to `describeIndex` and `deleteIndex` to both
set and delete the host from the cache when needed.

## Type of Change
- [X] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [X] This change requires a documentation update

## Test Plan
Testing is a bit tricky at the moment, but you should be able to pull
the branch down and test using the repl:

```
npm run repl

await init()
await client.listIndexes()

await client.index('my-index').describeIndexStats()
```

Make sure unit tests are passing in CI.

Integration tests are a bit borked right now, but there is a follow-up
ticket to address those specifically.