Add @spruce release to user-agent (#155)
_This revision is dependent on a previous branches / PRs. These will
need to be reviewed and merged first:_
- https://github.com/pinecone-io/pinecone-ts-client/pull/153
- https://github.com/pinecone-io/pinecone-ts-client/pull/154

We want to be able to track usage specific to the spruce release of the
client.

- Update `buildUserAgent` logic to check for `packageInfo.release`, and
append to the user agent if present.
- Update `release-spruce-dev` and `release-spruce` workflows to add
`@spruce` to `version.json` during these specific releases.

- [X] New feature (non-breaking change which adds functionality)

We will need to release through either the `release-spruce` or
`release-spruce-dev` workflows and validate the `version.json` file has
the proper info, and the `release` is appended to the user agent header.