Guard against empty hostUrl values in IndexHostSingleton cache (#165)
## Problem
We use `indexHostSingleton` to cache API keys and index `host` addresses
to avoid excessive `describeIndex` calls for determining an indexes
address. Currently, `IndexHostSingleton._set()` is able to push an empty
`''` `host` value into the cache.

If an index is created and described quickly, this could lead to a
situation where `host` is empty, which would break subsequent client
calls to that index.

## Solution
Guard against empty `host` values being passed to the cache.

- Throw a `PineconeUnableToResolveHostError` inside `_set` in the event
of an empty `hostUrl`
- Add unit test

## Type of Change
- [X] Bug fix (non-breaking change which fixes an issue)

## Test Plan
This one is tricky to test as you'd need a scenario where you call
`client.CreateIndex({ ... })` followed quickly by
`client.describeIndex('index')` before `host` is established. For now,
testing that the client works as expected should be sufficient.