[spruce] Update control plane operations (#167)
The schema for control plane operations has changed significantly. We
need to update all relevant methods, types, and runtime schemas to align
with the new client core.

For now the focus is on getting a working client, unit tests, TS
compilation, and everything other than integration tests passing.

- Updated all relevant control plane operations to support the new
request and response shapes.
- Key files to review are under `/src/control/*`. Tests and all of the
control modules. The primary differences are in the shape of request and
response objects.
- Updated the typebox runtime validation schemas for `configureIndex`
and `createIndex` to align with the new request shapes. There's some
jank in the errors that are thrown, we will need to make changes to
`validator.ts` (see `createIndex.validation.test.ts` or
`configureIndex.validation.test.ts`). For now the test files for the
validation on these 2 areas should be passing, I've added TODOs.
- Changes to integration tests were pulled in from my previous branch
and are mostly still relevant to get our builds passing. For now these
are still failing in CI and will be handled in a follow up.
- Added `additionalHeaders` as a field in `PineconeConfiguration`. This
is to support interacting with staging environments.

I've left some TODO comments throughout. I think there will need to be
additional passes and refactoring, but in terms of control plane and
data plane operations, this should support the new spec.

Key Follow-up items:
- Functional integration tests.
- Audit the generated types we're exporting. Confirm whether we want to
"put lipstick on the pig" in certain areas or leave things as-is.
- Update `validator.ts` to handle the new more nested shape of certain
arguments.
- Expand validation testing for creating configuring indexes.

- [X] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [X] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [X] This change requires a documentation update
- [X] Infrastructure change (CI configs, etc)
- [X] Non-code change (docs, etc)

To validate changes in staging, you'll need to pass a specific header in
via `PineconeConfiguration`.

Use the repl to validate control and data plane operations once the
correct configuration has been applied:

```
$ npm run repl
$ const client = new Pinecone({ apiKey: <your-api-key>, controllerHostUrl: <your-staging-url>, additionalHeaders: { 'x-environment': <your-dev-environment>}})

$ await client.listIndexes()
$ await client.createIndex({ .... })
...
```