[spruce] Update IndexHostSingleton with responses from listIndexes (#170)
## Problem
We currently use calls to `describeIndex` and `deleteIndex` to interact
with the `IndexHostSingleton` cache proactively. This allows bypassing
the need for the singleton to call `describeIndex` to retrieve the host
for data plane operations. We should be able to do the same thing with
`listIndexes` now that the response includes the full `IndexModel`.

## Solution

- Update `listIndexes` in the `Pinecone` class to call
`IndexHostSingleton._set(...)` for indexes that are listed.
- Add new unit test to cover behavior.

## Type of Change
- [X] New feature (non-breaking change which adds functionality)

## Test Plan
The new unit test should cover making sure we're calling
`IndexHostSingleton._set()` for `listIndexes` responses.

You can also pull this down and run locally by adding a `console.log()`
in `IndexHostSingleton._set()` to check the `hostUrls` map. You can use
the `PINECONE_DEBUG` and `PINECONE_DEBUG_CURL` environment variables to
verify that calling `listIndexes()` before making dataplane operations
results in no unnecessary calls to `describeIndex` by the cache.