Move non-runtime related deps to devDependencies (#175)
## Problem

The purpose of this change is to significantly reduce the size of the
package when installed. The issue is fully described in #174.

## Solution

The solution is to move the non-runtime related dependencies into the
`devDependencies` section of the `package.json` file so that npm doesn't
try to install them in a nested `node_modules` folder.

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [x] None of the above: dependency tweaks

## Test Plan

The code compiled and the unit checks passed. The packages moved are
either entirely type declaration files (i.e., `<name>.d.ts`) and/or
needed only for compilation (and would be likely be included in the
parent package anyway if actually needed). That said, if there's any
doubts, we'd know for sure if it'd work after pruning the dev
dependencies and trying it out.